Fix: Apple login
