+++ hoohoo_web/src/Pages/Login/pages/Login.tsx
@@ -368,7 +368,7 @@ const Login = () => {
                     <GoogleLogo text={localizedTexts.continueWithGoogle} />
                   </SocialButton>
 
-                  {/* <SocialButton onClick={handleAppleLogin}>
+                  <SocialButton onClick={handleAppleLogin}>
                     <SocialInnerBox>
                       <svg
                         width="24"
@@ -382,7 +382,7 @@ const Login = () => {
                       </svg>
                       <span>{localizedTexts.continueWithApple}</span>
                     </SocialInnerBox>
-                  </SocialButton> */}
+                  </SocialButton>
 
                   <KakaoButton onClick={handleKakaoLogin}>
                     <SocialInnerBox>
