+++ hoohoo_web/src/Pages/Login/pages/Signup.tsx
@@ -309,7 +309,7 @@ const Signup = () => {
               <GoogleLogo text={localizedTexts.signupWithGoogle} />
             </SocialButton>
 
-            {/* <SocialButton onClick={handleAppleLogin}>
+            <SocialButton onClick={handleAppleLogin}>
               <SocialInnerBox>
                 <svg
                   width="24"
@@ -323,7 +323,7 @@ const Signup = () => {
                 </svg>
                 <span>{localizedTexts.signupWithApple}</span>
               </SocialInnerBox>
-            </SocialButton> */}
+            </SocialButton>
 
             <KakaoButton onClick={handleKakaoLogin}>
               <SocialInnerBox>
